scapi could not compile on osx mavericks before since in mavericks apple moved to the clang compiler (based on llvm) whether the makefile is built on the assumption that the compiler is gcc.
in order to maintain a unified makefile for both linux and mac os x,
there is a need to manually specify gcc-4.9 as the compiler (only for mavericks users).

this is now handled automatically in the makefile, but the user is required to install
gcc49 via homebrew. this will be addressed in the online documentaion.